pyproject.toml
[project]
name = "general-contractor-agent-demo"
version = "0.1.0"
description = "Multi-agent orchestration system demonstrating construction project management with AI agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.34.0",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "strands-agents>=0.5.0",
    "uvicorn>=0.38.0",
]
